Added an eventlistener that listens for mouseover events.
This eventlistener is applied to the container div which houses all the divs that make up the grid.

Once activated, the specific grid div would have its background color set to red through the classList.toggle method.

Included a set timer within the eventlistener function that returns the background color to white after 500 ms.

Currently the eventlistener is not working.